# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InfraPulse MSK",
    "description": "API InfraPulse MSK. Документация открыта для чтения; каждый маршрут, кроме `/health/*` и входа, проверяет сессию или токен интеграции.\n\n* **Чтение (GET)** — по cookie сессии: войти в веб-интерфейс этого же сайта, затем «Try it out» здесь.\n* **Пачки наблюдений** `POST /api/v1/observations` — токен интеграции: «Authorize» → `integrationToken`, значение `ipk_...` без слова Bearer.\n* **Запись под сессией (POST)** — дополнительно `csrfToken`: значение cookie `__Host-infrapulse-csrf`.\n\nИнструкция для систем-источников — `docs/INTEGRATION_API.md` в репозитории.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/forecasts/{forecast_id}/decisions": {
      "get": {
        "summary": "Decisions",
        "operationId": "decisions_api_v1_forecasts__forecast_id__decisions_get",
        "parameters": [
          {
            "name": "forecast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Forecast Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastDecisionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "summary": "Decide",
        "operationId": "decide_api_v1_forecasts__forecast_id__decisions_post",
        "parameters": [
          {
            "name": "forecast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Forecast Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastDecisionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastDecisionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
    "/api/v1/forecasts/{forecast_id}/check-results": {
      "get": {
        "summary": "Check Results",
        "operationId": "check_results_api_v1_forecasts__forecast_id__check_results_get",
        "parameters": [
          {
            "name": "forecast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Forecast Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastCheckResultList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/forecasts/{forecast_id}/check-result": {
      "post": {
        "summary": "Record Check Result",
        "operationId": "record_check_result_api_v1_forecasts__forecast_id__check_result_post",
        "parameters": [
          {
            "name": "forecast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Forecast Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastCheckResultCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastCheckResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForecastCheckResult": {
        "properties": {
          "forecast_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Forecast Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "check_result": {
            "type": "string",
            "enum": [
              "awaiting",
              "fixed",
              "no_violation",
              "not_done"
            ],
            "title": "Check Result"
          },
          "found": {
            "items": {
              "type": "string",
              "enum": [
                "breaker",
                "cable",
                "contactor",
                "comm_module",
                "cabinet_power",
                "other"
              ]
            },
            "type": "array",
            "title": "Found"
          },
          "found_other_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Found Other Text"
          },
          "found_confirmed_by_customer": {
            "type": "boolean",
            "const": false,
            "title": "Found Confirmed By Customer",
            "default": false
          },
          "result_at": {
            "type": "string",
            "format": "date-time",
            "title": "Result At"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "event_cause": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "planned_outage",
                  "protection_trip",
                  "external_grid",
                  "smvu_channel",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Cause"
          },
          "actor_id": {
            "type": "string",
            "minLength": 1,
            "title": "Actor Id"
          },
          "actor_role": {
            "type": "string",
            "minLength": 1,
            "title": "Actor Role"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "forecast_id",
          "revision",
          "check_result",
          "result_at",
          "actor_id",
          "actor_role",
          "recorded_at",
          "simulated"
        ],
        "title": "ForecastCheckResult",
        "description": "Stored check result revision of a card (append-only, audited)."
      },
      "ForecastCheckResultCreate": {
        "properties": {
          "check_result": {
            "type": "string",
            "enum": [
              "awaiting",
              "fixed",
              "no_violation",
              "not_done"
            ],
            "title": "Check Result"
          },
          "found": {
            "items": {
              "type": "string",
              "enum": [
                "breaker",
                "cable",
                "contactor",
                "comm_module",
                "cabinet_power",
                "other"
              ]
            },
            "type": "array",
            "title": "Found"
          },
          "found_other_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Found Other Text"
          },
          "result_at": {
            "type": "string",
            "format": "date-time",
            "title": "Result At"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "event_cause": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "planned_outage",
                  "protection_trip",
                  "external_grid",
                  "smvu_channel",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Cause"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "check_result",
          "result_at",
          "idempotency_key",
          "expected_revision"
        ],
        "title": "ForecastCheckResultCreate",
        "description": "Result of the check after a decision (C0.4); the author comes from the session.\n\n``found`` («что устранено») only with ``fixed`` and is not confirmed by the customer;\n``event_cause`` only for a card whose event was registered (the server checks it)."
      },
      "ForecastCheckResultList": {
        "properties": {
          "forecast_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Forecast Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ForecastCheckResult"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "forecast_id",
          "items"
        ],
        "title": "ForecastCheckResultList",
        "description": "Every check result revision of one card, newest first."
      },
      "ForecastDecisionCreate": {
        "properties": {
          "decision_code": {
            "type": "string",
            "enum": [
              "R1",
              "R2",
              "R3",
              "R4",
              "R5",
              "R6",
              "R7"
            ],
            "title": "Decision Code"
          },
          "reason_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Reason Code"
          },
          "reason_text": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Reason Text"
          },
          "verification_methods": {
            "items": {
              "type": "string",
              "enum": [
                "source_records",
                "remote_poll",
                "call_collector",
                "video",
                "field_visit",
                "not_checked"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Verification Methods"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          },
          "draft_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Note"
          },
          "notified_to": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified To"
          },
          "notified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified At"
          },
          "awaiting_result_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awaiting Result Until"
          },
          "watch_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Until"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision_code",
          "reason_code",
          "reason_text",
          "verification_methods",
          "idempotency_key",
          "expected_revision"
        ],
        "title": "ForecastDecisionCreate",
        "description": "Dispatcher decision request; the author comes from the session, never the body.\n\n``idempotency_key`` repeats return the stored decision; ``expected_revision`` is the\ncard's decision revision the dispatcher saw (0 when none), a mismatch is 409."
      },
      "ForecastDecisionList": {
        "properties": {
          "forecast_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Forecast Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ForecastDecisionSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "forecast_id",
          "items"
        ],
        "title": "ForecastDecisionList",
        "description": "Every decision revision of one card, newest first, with its author."
      },
      "ForecastDecisionSummary": {
        "properties": {
          "decision_code": {
            "type": "string",
            "enum": [
              "R1",
              "R2",
              "R3",
              "R4",
              "R5",
              "R6",
              "R7"
            ],
            "title": "Decision Code"
          },
          "reason_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Reason Code"
          },
          "reason_text": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Reason Text"
          },
          "dictionary_version": {
            "type": "string",
            "minLength": 1,
            "title": "Dictionary Version"
          },
          "actor_id": {
            "type": "string",
            "minLength": 1,
            "title": "Actor Id"
          },
          "actor_role": {
            "type": "string",
            "minLength": 1,
            "title": "Actor Role"
          },
          "decided_at": {
            "type": "string",
            "format": "date-time",
            "title": "Decided At"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "verification_methods": {
            "items": {
              "type": "string",
              "enum": [
                "source_records",
                "remote_poll",
                "call_collector",
                "video",
                "field_visit",
                "not_checked"
              ]
            },
            "type": "array",
            "title": "Verification Methods"
          },
          "notified_to": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified To"
          },
          "notified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified At"
          },
          "awaiting_result_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awaiting Result Until"
          },
          "watch_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch Until"
          },
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "draft_status": {
            "anyOf": [
              {
                "type": "string",
                "const": "not_sent"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Status"
          },
          "check_result_code": {
            "type": "string",
            "enum": [
              "O1",
              "O2",
              "O3",
              "O4",
              "O5",
              "O6"
            ],
            "title": "Check Result Code",
            "default": "O6"
          },
          "check_result_simulated": {
            "type": "boolean",
            "title": "Check Result Simulated",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision_code",
          "reason_code",
          "reason_text",
          "dictionary_version",
          "actor_id",
          "actor_role",
          "decided_at",
          "revision",
          "simulated"
        ],
        "title": "ForecastDecisionSummary",
        "description": "Dispatcher decision (technologist dictionary R1–R7) and later check result (O1–O6)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "integrationToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ipk_...",
        "description": "Токен интеграции внешней системы (выдаёт администратор сервиса). Даёт только право ingest: отправку пачек и статус своих пачек."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-infrapulse-session",
        "description": "Сессия входа через каталог. Ставится при входе в веб-интерфейс этого же сайта и отправляется браузером сама; поле в «Authorize» заполнять не нужно."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Для POST под сессией: значение cookie __Host-infrapulse-csrf. С токеном интеграции не нужен."
      }
    }
  }
}